VISTA-553 bedrooms and bathrooms are now integers

+++ vista-portal/vista-portal-domain/src/main/java/com/propertyvista/portal/domain/dto/FloorplanPropertyDTO.java
@@ -26,7 +26,7 @@ public interface FloorplanPropertyDTO extends IEntity {
 
     RangeGroup price();
 
-    IPrimitive<Double> bedrooms();
+    IPrimitive<Integer> bedrooms();
 
-    IPrimitive<Double> bathrooms();
+    IPrimitive<Integer> bathrooms();
 }
